StatusContainer created
